Added menu, nonLinearCave fixed, exits added
A simple button object and menu setup were created with an event queue.
Flow structure was added to allow switching between game and menu. Fixed
nonLinearCave to allow it to be added. Exit stairs are now being added
in a simple manner. collision with stairs causes the player to advance
to the next level.