Adjusted collision detection scaling values.
Scaled corner values by values for player to adjust the size of the
hitbox. Find a different and more robust solution in the future.
Structured ai.lua, template basic.lua ai file created, filled in the
remainder of enemy.lua, and added draw call and event processing for
enemy to world.lua